+++ collector-bitmovin-player/src/androidTest/java/com/bitmovin/analytics/bitmovin/player/CronetNetworkTest.kt
@@ -1,6 +1,5 @@
 package com.bitmovin.analytics.bitmovin.player
 
-import android.content.Context
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import com.bitmovin.analytics.api.AnalyticsConfig
@@ -15,14 +14,12 @@ import com.bitmovin.analytics.systemtest.utils.runBlockingTest
 import com.bitmovin.player.api.PlaybackConfig
 import com.bitmovin.player.api.Player
 import com.bitmovin.player.api.PlayerConfig
-import com.bitmovin.player.api.TweaksConfig
 import com.bitmovin.player.api.analytics.create
+import com.bitmovin.player.api.source.NetworkEngine
 import com.bitmovin.player.api.source.Source
 import com.bitmovin.player.api.source.SourceConfig
-import com.google.android.gms.net.CronetProviderInstaller
-import kotlinx.coroutines.CompletableDeferred
+import com.bitmovin.player.api.source.SourceNetworkConfig
 import kotlinx.coroutines.MainScope
-import kotlinx.coroutines.future.asCompletableFuture
 import kotlinx.coroutines.withContext
 import org.assertj.core.api.Assertions.assertThat
 import org.chromium.net.CronetEngine
@@ -31,7 +28,6 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import java.util.concurrent.Future
 
 @RunWith(AndroidJUnit4::class)
 class CronetNetworkTest {
@@ -49,11 +45,6 @@ class CronetNetworkTest {
         // Unused setter
         set(_) {}
 
-    // Source depends on defaultSourceMetaData which depends on the Test, so it has to be generated dynamically
-    private var defaultSource: Source
-        get() = Source.create(SourceConfig.fromUrl(defaultSample.m3u8Url!!), defaultSourceMetadata)
-        set(_) {}
-
     private lateinit var defaultAnalyticsConfig: AnalyticsConfig
     private lateinit var mockedIngressUrl: String
     private lateinit var defaultPlayer: Player
@@ -62,7 +53,6 @@ class CronetNetworkTest {
         PlayerConfig(
             key = "a6e31908-550a-4f75-b4bc-a9d89880a733",
             playbackConfig = PlaybackConfig(isMuted = true),
-            tweaksConfig = TweaksConfig(cronetEngine = loadCronet(appContext)),
         )
 
     @Before
@@ -90,8 +80,18 @@ class CronetNetworkTest {
     @Test
     fun test_downloadTracking_withCronetNetworkStack() =
         runBlockingTest {
+            val sourceConfig = SourceConfig.fromUrl(defaultSample.m3u8Url!!)
+            sourceConfig.networkConfig =
+                SourceNetworkConfig(
+                    engine =
+                        NetworkEngine.Cronet(
+                            CronetEngine.Builder(appContext).build(),
+                        ),
+                )
+
+            val source = Source.create(sourceConfig, defaultSourceMetadata)
             withContext(mainScope.coroutineContext) {
-                defaultPlayer.load(defaultSource)
+                defaultPlayer.load(source)
                 defaultPlayer.play()
             }
 
@@ -111,13 +111,4 @@ class CronetNetworkTest {
             val eventDataList = impression.eventDataList
             DataVerifier.verifyBandwidthMetrics(eventDataList)
         }
-
-    private fun loadCronet(appContext: Context): Future<CronetEngine?> {
-        // See https://developer.android.com/codelabs/cronet#5
-        val result = CompletableDeferred<CronetEngine?>()
-        CronetProviderInstaller.installProvider(appContext).addOnCompleteListener {
-            result.complete(CronetEngine.Builder(appContext).build())
-        }
-        return result.asCompletableFuture()
-    }
 }